backend/foodgram/requirements.txt
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
sqlparse==0.3.1
pytz==2020.1
psycopg2-binary==2.8.6
pillow==9.4.0
djoser==2.1.0
drf-base64==2.0
flake8==5.0.4
gunicorn==20.1.0
pytest-pythonpath==0.7.3
pytest==6.2.5
pytest-django==4.4.0